requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==5.1.1
django-filter==24.3
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
dostoevsky==0.6.0
et-xmlfile==1.1.0
fasttext==0.9.2
idna==3.10
joblib==1.4.2
mysqlclient==2.2.4
numpy==2.1.1
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.2.3
pyaes==1.6.1
pyasn1==0.6.1
pybind11==2.13.6
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.2
razdel==0.5.0
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
rsa==4.9
setuptools==75.1.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.5.1
Telethon==1.37.0
tqdm==4.66.5
tzdata==2024.2
urllib3==2.2.3
wheel==0.44.0
